requirements.txt
appdirs==1.4.4
asgiref==3.7.2
baron==0.10.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
dj-email-url==1.0.6
Django==5.0.3
django-cache-url==3.4.5
django-cors-headers==4.3.1
django-environ==0.11.2
django-environs==0.1.1
django-phonenumber-field==7.3.0
django-user-sessions==2.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
environs==11.0.0
idna==3.6
marshmallow==3.21.1
packaging==23.2
phonenumbers==8.13.31
pillow==10.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
redbaron==0.9.2
requests==2.31.0
rply==0.7.8
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
